requirements.txt
google-generativeai==0.8.3
PyGithub==2.5.0
gitpython==3.1.43
pydantic==2.10.3
python-dotenv==1.0.1
rich==13.9.4
pyyaml==6.0.2
requests==2.32.3